requirements.txt
# Flask 核心
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7

# 数据库
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7

# Redis
redis==5.0.1

# 异步支持
eventlet==0.33.3
greenlet==3.0.1
python-socketio==5.10.0
python-engineio==4.9.0

# 工具
PyJWT==2.8.0
bcrypt==4.1.1
python-dotenv==1.0.0

# 其他
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
